- updated test project to netcoreapp2.1 - new byte array comparison for correct results in linux - added a native test to check correct type sizes on linux
git-svn-id: file:///H/Users/C_B-A/temp/NCurses.Core_git/branches/span@38 3a9065fc-4a7e-7541-9dc8-748989605d84

+++ src/NCurses.Core/Interop/MultiByteString/NativeStdScrWideStr.cs
@@ -214,7 +214,7 @@ namespace NCurses.Core.Interop.MultiByteString
             unsafe
             {
                 TWideStr* strPtr = stackalloc TWideStr[n];
-                read = NCursesException.Verify((int)this.Wrapper.mvinnwstr(y, x, ref MarshalString(strPtr, n, out Span<TWideStr> span), n), "mvinnwstr");
+                read = NCursesException.Verify(this.Wrapper.mvinnwstr(y, x, ref MarshalString(strPtr, n, out Span<TWideStr> span), n), "mvinnwstr");
                 wstr = ReadString(ref span);
             }
         }
